move state package (checkpoint)
