+++ lib/streamlit/components/v1/components.py
@@ -27,7 +27,7 @@ from streamlit.logger import get_logger
 from streamlit.proto.Components_pb2 import SpecialArg, ArrowTable as ArrowTableProto
 from streamlit.proto.Element_pb2 import Element
 from streamlit.runtime.scriptrunner import get_script_run_ctx
-from streamlit.state import NoValue, register_widget
+from streamlit.runtime.state import NoValue, register_widget
 from streamlit.type_util import to_bytes
 
 LOGGER = get_logger(__name__)
