+++ lib/streamlit/elements/number_input.py
@@ -22,7 +22,7 @@ import streamlit
 from streamlit.errors import StreamlitAPIException
 from streamlit.js_number import JSNumber, JSNumberBoundsException
 from streamlit.proto.NumberInput_pb2 import NumberInput as NumberInputProto
-from streamlit.state import (
+from streamlit.runtime.state import (
     register_widget,
     NoValue,
     WidgetArgs,
