+++ lib/streamlit/elements/utils.py
@@ -19,7 +19,7 @@ import streamlit
 from streamlit import type_util
 from streamlit.elements.form import is_in_form
 from streamlit.errors import StreamlitAPIException
-from streamlit.state import get_session_state, WidgetCallback
+from streamlit.runtime.state import get_session_state, WidgetCallback
 
 if TYPE_CHECKING:
     from streamlit.delta_generator import DeltaGenerator
