+++ lib/tests/streamlit/delta_generator_test.py
@@ -31,8 +31,8 @@ from streamlit.proto.TextInput_pb2 import TextInput
 from streamlit.proto.Empty_pb2 import Empty as EmptyProto
 from streamlit.proto.RootContainer_pb2 import RootContainer
 from streamlit.proto.Text_pb2 import Text as TextProto
-from streamlit.state.widgets import _build_duplicate_widget_message
-import streamlit.state.widgets as w
+from streamlit.runtime.state.widgets import _build_duplicate_widget_message
+import streamlit.runtime.state.widgets as w
 from tests import testutil
