+++ lib/tests/streamlit/script_run_context_test.py
@@ -17,7 +17,7 @@ import unittest
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.runtime.scriptrunner import ScriptRunContext
-from streamlit.state import SafeSessionState, SessionState
+from streamlit.runtime.state import SafeSessionState, SessionState
 from streamlit.runtime.uploaded_file_manager import UploadedFileManager
