+++ lib/tests/streamlit/state/safe_session_state_test.py
@@ -16,7 +16,7 @@ import unittest
 from typing import Tuple, Dict, Any
 from unittest.mock import MagicMock, Mock, PropertyMock
 
-from streamlit.state import SessionState, SafeSessionState
+from streamlit.runtime.state import SessionState, SafeSessionState
 
 
 def _create_state_spy(
