+++ lib/tests/streamlit/state/session_state_proxy_test.py
@@ -21,11 +21,9 @@ from unittest.mock import patch
 import pytest
 
 from streamlit.errors import StreamlitAPIException
-from streamlit.state import SafeSessionState
-from streamlit.state.session_state_proxy import SessionStateProxy
-from streamlit.state.session_state import (
+from streamlit.runtime.state import SafeSessionState, SessionState, SessionStateProxy
+from streamlit.runtime.state.session_state import (
     GENERATED_WIDGET_KEY_PREFIX,
-    SessionState,
     require_valid_user_key,
 )
 
@@ -43,7 +41,7 @@ def _create_mock_session_state(
 
 
 @patch(
-    "streamlit.state.session_state_proxy.get_session_state",
+    "streamlit.runtime.state.session_state_proxy.get_session_state",
     return_value=_create_mock_session_state({"foo": "bar"}),
 )
 class SessionStateProxyTests(unittest.TestCase):
@@ -107,7 +105,7 @@ class SessionStateProxyAttributeTests(unittest.TestCase):
         self.session_state_proxy = SessionStateProxy()
 
     @patch(
-        "streamlit.state.session_state_proxy.get_session_state",
+        "streamlit.runtime.state.session_state_proxy.get_session_state",
         return_value=SessionState(_new_session_state={"foo": "bar"}),
     )
     def test_delattr(self, _):
@@ -115,14 +113,14 @@ class SessionStateProxyAttributeTests(unittest.TestCase):
         assert "foo" not in self.session_state_proxy
 
     @patch(
-        "streamlit.state.session_state_proxy.get_session_state",
+        "streamlit.runtime.state.session_state_proxy.get_session_state",
         return_value=SessionState(_new_session_state={"foo": "bar"}),
     )
     def test_getattr(self, _):
         assert self.session_state_proxy.foo == "bar"
 
     @patch(
-        "streamlit.state.session_state_proxy.get_session_state",
+        "streamlit.runtime.state.session_state_proxy.get_session_state",
         return_value=SessionState(_new_session_state={"foo": "bar"}),
     )
     def test_getattr_error(self, _):
@@ -130,7 +128,7 @@ class SessionStateProxyAttributeTests(unittest.TestCase):
             del self.session_state_proxy.nonexistent
 
     @patch(
-        "streamlit.state.session_state_proxy.get_session_state",
+        "streamlit.runtime.state.session_state_proxy.get_session_state",
         return_value=SessionState(_new_session_state={"foo": "bar"}),
     )
     def test_setattr(self, _):
