+++ lib/tests/streamlit/state/strategies.py
@@ -12,12 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from streamlit.state.session_state import (
-    SessionState,
+from hypothesis import strategies as hst
+
+from streamlit.runtime.state.session_state import (
     GENERATED_WIDGET_KEY_PREFIX,
     WidgetMetadata,
+    SessionState,
 )
-from hypothesis import strategies as hst
 
 ASCII = list("abcdefghijklmnopqrstuvwxyz0123456789_-")
